requirements.txt


# Core dependencies
PyYAML
requests
urllib3
colorama
# Async & HTTP
aiohttp
# Data parsing, reporting, and utilities
lxml
beautifulsoup4
jsonschema
regex

# AI Providers (install if AI modules enabled or planning to use Groq, OpenAI, etc.)
openai
httpx

# Environmental variable support
python-dotenv

# JWT, token, and auth handling
PyJWT

# Crawling / Browser Automation (only if you enable JS crawling)
playwright
selenium
# Fast API server (for any OAST client/server use)
fastapi
uvicorn

# Optional: CLI frameworks
click
# Data analysis for advanced reports/statistics
numpy
pandas

# Packaging and setup helper
setuptools

# Type hints
typing-extensions
# Testing and linting (optional, for development only)
pytest
flake8
black
